Move more methods to ReadBuffer
