+++ Aaru.Devices/Device/ScsiCommands/LiteOn.cs
@@ -406,69 +406,4 @@ public partial class Device
             _ => DeinterleaveFullEccInterleaved(buffer, transferLength, stride) // Default fallback
         };
     }
-
-    /// <summary>
-    ///     Deinterleave the ECC block stored within a raw sector (backward compatibility wrapper)
-    /// </summary>
-    /// <param name="buffer">Data buffer</param>
-    /// <param name="transferLength">How many blocks in buffer</param>
-    /// <returns>The deinterleaved sectors</returns>
-    private byte[] DeinterleaveEccBlock(byte[] buffer, uint transferLength)
-    {
-        return DeinterleaveEccBlock(buffer, transferLength, _bufferStride, _bufferFormat);
-    }
-
-    /// <summary>
-    ///     Deinterleave full ECC block with interleaved PI (e.g., 2384 bytes)
-    /// </summary>
-    /// <param name="buffer">Data buffer</param>
-    /// <param name="transferLength">How many blocks in buffer</param>
-    /// <param name="stride">Bytes per sector in buffer</param>
-    /// <returns>The deinterleaved sectors</returns>
-    private static byte[] DeinterleaveFullEccInterleaved(byte[] buffer, uint transferLength, uint stride)
-    {
-        // TODO: Save ECC instead of just throwing it away
-
-        var deinterleaved = new byte[2064 * transferLength];
-
-        for(var j = 0; j < transferLength; j++)
-        {
-            for(var i = 0; i < 12; i++) Array.Copy(buffer, j * stride + i * 182, deinterleaved, j * 2064 + i * 172, 172);
-        }
-
-        return deinterleaved;
-    }
-
-    /// <summary>
-    ///     Extract sector data from PO-only format (e.g., 2236 bytes)
-    /// </summary>
-    /// <param name="buffer">Data buffer</param>
-    /// <param name="transferLength">How many blocks in buffer</param>
-    /// <param name="stride">Bytes per sector in buffer</param>
-    /// <returns>The extracted sector data</returns>
-    private static byte[] DeinterleavePoOnly(byte[] buffer, uint transferLength, uint stride)
-    {
-        var deinterleaved = new byte[2064 * transferLength];
-
-        for(var j = 0; j < transferLength; j++)
-        {
-            Array.Copy(buffer, j * stride, deinterleaved, j * 2064, 2064);
-        }
-
-        return deinterleaved;
-    }
-
-    /// <summary>
-    ///     Deinterleave full ECC block with padding (e.g., 2816 bytes)
-    /// </summary>
-    /// <param name="buffer">Data buffer</param>
-    /// <param name="transferLength">How many blocks in buffer</param>
-    /// <param name="stride">Bytes per sector in buffer</param>
-    /// <returns>The deinterleaved sectors</returns>
-    private static byte[] DeinterleaveFullEccWithPadding(byte[] buffer, uint transferLength, uint stride)
-    {
-        // Same as FullEccInterleaved, padding is ignored
-        return DeinterleaveFullEccInterleaved(buffer, transferLength, stride);
-    }
-
 }
